Move code from Main to Lib
It's needed to enable testing, because tests can't import app/Main.hs
