+++ src/Lib.hs
@@ -1,6 +1,70 @@
-module Lib
-    ( someFunc
-    ) where
+module Lib where
+
+type Identifier = String
+
+type Env = [(Identifier, Value)]
+
+data Expr
+  = Lit String
+  | Term Identifier
+  | Abs Identifier Expr
+  | App Expr Expr
+  deriving (Show)
+
+l_id :: Expr
+l_id = Abs "x" (Term "x")
+
+l_const :: Expr
+l_const = Abs "x" (Abs "y" (Term "x"))
+
+test_1 :: Expr
+test_1 = App l_id (Lit "wow")
+
+test_2 :: Expr
+test_2 = App (App l_const (Lit "*")) (Lit "!")
+
+test_3 :: Expr
+test_3 =
+  App
+    (App (App l_const l_id) (Lit "!"))
+    (App (App l_const (Lit "*")) l_id)
+
+test_4 :: Expr
+test_4 =
+  App l_id (App l_const (Lit "*"))
+
+test_err_1 :: Expr
+test_err_1 =
+  App (App l_id (Lit "*")) (Lit "!")
+
+data Value
+  = Value String
+  | Closure Expr Env Identifier
+  deriving (Show)
+
+eval :: Env -> Expr -> Maybe Value
+eval env (Lit string)          = Just $ Value string
+eval env (Term identifier)     = lookup identifier env
+eval env (Abs identifier expr) = Just $ Closure expr env identifier
+eval env (App t u) =
+  let et = eval env t
+      eu = eval env u
+   in case et of
+        Just (Closure expr env' arg) ->
+          case eu of
+            Nothing    -> Nothing
+            Just value -> eval ((arg, value) : env') expr
+        _                            -> Nothing
+
+checkShadowing :: [Identifier] -> Expr -> [Identifier]
+checkShadowing args (Abs arg expr) 
+  | arg !! 0 == '_' = nested
+  | elem arg args   = arg : nested
+  | otherwise       = nested
+    where 
+      nested  = checkShadowing (arg : args) expr
+checkShadowing args (App t u) = concatMap (checkShadowing args) [t, u]
+checkShadowing _ _ = []
 
 someFunc :: IO ()
 someFunc = putStrLn "someFunc"
